fix homework-2
