+++ homework-2/conftest.py
@@ -33,7 +33,8 @@ def pytest_configure(config):
 
 @pytest.fixture(scope='function')
 def test_dir(request):
-    test_dir = os.path.join(request.config.base_test_dir, request._pyfuncitem.nodeid)
+    test_name = request._pyfuncitem.nodeid.replace('/', '_').replace(':', '_')
+    test_dir = os.path.join(request.config.base_test_dir, test_name)
     os.makedirs(test_dir)
     return test_dir
