+++ homework-2/locators/top_panel_locators.py
@@ -1,3 +1,5 @@
+from selenium.webdriver.common.by import By
+
 # Top panel module names
 tags = ['SEGMENTS', 'BILLING', 'STATISTICS','PRO',
  'PROFILE', 'TOOLS', 'DASHBOARD', 'LOGOUT']
@@ -14,3 +16,8 @@ urls = [
 'https://target.my.com/profile/contacts',
 'https://target.my.com/tools/feeds'
 ]
+
+
+class TopPanelLocators:
+    RIGHT_MODULE_LOCATOR = (By.XPATH, '//*[starts-with(@class, "right-module-rightButton")]')
+    BUTTONS_LOCATORS = {tags[i]: (By.XPATH, href[i]) for i in range(len(tags))}
